+++ src/components/JoinPage/common/Header/style.ts
