+++ src/components/JoinPage/common/PageProgressBar/style.ts
